+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.5 - 2020-10-22
+
+### Fix
+- Posts entity change name Docs
+- If there is no data, a 404 redirect is for the homepage
+- admin/docs width medium to width large for sidebar
+
 ## 1.0.3 - 2020-09-05
 
 ### Fix